+++ src/cjs/utils/datastructures/GenericPath.js
@@ -38,23 +38,23 @@ var GenericPath = /** @class */ (function () {
      * @return {Path} A deep clone/copy of this path.
      */
     GenericPath.prototype.clone = function () {
-        var newPath = new GenericPath(this.segments[this.segments.length - 1].clone().revert());
+        var newPath = new GenericPath(this.segments[this.segments.length - 1].clone().reverse());
         for (var i = this.segments.length - 2; i >= 0; i--) {
-            newPath.segments.push(this.segments[i].clone().revert());
+            newPath.segments.push(this.segments[i].clone().reverse());
         }
         return newPath;
     };
     /**
      * Reverse this path (swap start and end and thus – the direction) in-place.
      *
-     * @method revert
+     * @method reverse
      * @memberof Path
      * @return {PathSegment} This path instance.
      */
-    GenericPath.prototype.revert = function () {
+    GenericPath.prototype.reverse = function () {
         var newSegments = [];
         for (var i = this.segments.length - 1; i >= 0; i--) {
-            newSegments.push(this.segments[i].revert());
+            newSegments.push(this.segments[i].reverse());
         }
         this.segments = newSegments;
         return this;
@@ -78,6 +78,26 @@ var GenericPath = /** @class */ (function () {
      */
     GenericPath.prototype.getEndPoint = function () {
         return this.segments[this.segments.length - 1].getEndPoint();
+    };
+    /**
+     * Get the tangent's end point at the start point of this segment.
+     *
+     * @method getStartTangent
+     * @memberof PathSegment
+     * @return {Vertex} The end point of the starting point's tangent.
+     */
+    GenericPath.prototype.getStartTangent = function () {
+        return this.segments[0].getStartTangent();
+    };
+    /**
+     * Get the tangent's end point at the end point of this segment.
+     *
+     * @method getEndTangent
+     * @memberof PathSegment
+     * @return {Vertex} The end point of the ending point's tangent.
+     */
+    GenericPath.prototype.getEndTangent = function () {
+        return this.segments[this.segments.length - 1].getEndTangent();
     };
     return GenericPath;
 }());