+++ src/ts/CubicBezierCurve.ts
@@ -20,8 +20,8 @@
  * @modified 2021-01-20 Added UID.
  * @modified 2022-02-02 Added the `destroy` method.
  * @modified 2022-02-02 Cleared the `toSVGPathData` function (deprecated). Use `drawutilssvg` instead.
- * @modified 2022-10-17 Addint these method from the `PathSegment` interface: revert.
- * @version 2.7.0
+ * @modified 2022-10-17 The `CubicBezierCurve` class now implements the new `PathSegment` interface.
+ * @version 2.7.1
  *
  * @file CubicBezierCurve
  * @public
@@ -628,18 +628,29 @@ export class CubicBezierCurve implements PathSegment {
     );
   }
 
-  // TODO: docs?
-  revert(): CubicBezierCurve {
-    const sp = this.startPoint;
-    const scp = this.startControlPoint;
-    const ep = this.endPoint;
-    const ecp = this.endControlPoint;
-    this.startPoint = ep;
-    this.startControlPoint = ecp;
-    this.endPoint = sp;
-    this.endControlPoint = scp;
-    return this;
+  //---BEGIN PathSegment-------------------------
+  /**
+   * Get the tangent's end point at the start point of this segment.
+   *
+   * @method getStartTangent
+   * @memberof PathSegment
+   * @return {Vertex} The end point of the starting point's tangent.
+   */
+  getStartTangent(): Vertex {
+    return this.startControlPoint;
   }
+
+  /**
+   * Get the tangent's end point at the end point of this segment.
+   *
+   * @method getEndTangent
+   * @memberof PathSegment
+   * @return {Vertex} The end point of the ending point's tangent.
+   */
+  getEndTangent(): Vertex {
+    return this.endControlPoint;
+  }
+  //---END PathSegment-------------------------
 
   /**
    * Check if this and the specified curve are equal.<br>