+++ src/ts/Line.ts
@@ -110,21 +110,39 @@ export class Line extends VertTuple<Line> implements SVGSerializable, PathSegmen
   }
 
   /**
-   * Create a deep clone of this path segment.
+   * Get the tangent's end point at the start point of this segment.
    *
-   * @method clone
+   * @method getStartTangent
    * @memberof PathSegment
-   * @return {PathSegment} A deep clone/copy of this path segment.
+   * @return {Vertex} The end point of the starting point's tangent.
    */
-  //  clone: () => PathSegment;
+  getStartTangent(): Vertex {
+    return this.b;
+  }
+
+  /**
+   * Get the tangent's end point at the end point of this segment.
+   *
+   * @method getEndTangent
+   * @memberof PathSegment
+   * @return {Vertex} The end point of the ending point's tangent.
+   */
+  getEndTangent(): Vertex {
+    return this.a;
+  }
 
   /**
    * Inverse this path segment (in-place) and return this same instance (useful for chaining).
    *
-   * @method revert
+   * @method reverse
    * @memberof PathSegment
    * @return {PathSegment} This path segment instance (for chaining).
    */
-  revert: () => PathSegment;
+  reverse(): Line {
+    var tmp = this.a;
+    this.a = this.b;
+    this.b = tmp;
+    return this;
+  }
   //--- END Implement PathSegment ---
 }