+++ changelog.md
@@ -1,10 +1,10 @@
 # Changelog
 
 - 2022-10-17
+  - The `CubicBezierCurve` class now implements the new `PathSegment` interface.
   - Adding the `algorithms/detectPaths` function.
   - Changing the return type of `GirihTile.transformTilePositionToAdjacency` from `Polygon`to `Polygon | null`.
-  - Adding these methods from the `PathSegment` interface to the `Line` class: getStartPoint, getEndPoint, revert.
-  - Adding these methods from the `PathSegment` interface to the `CubicBezierCurve` class: revert.
+  - Adding these methods from the `PathSegment` interface to the `Line` class: getStartPoint, getEndPoint, reverse.
 - 2022-10-07
   - Adding the 44-truchet-tiles demo.
   - Adding the `algorithms/arrayShuffle` function.